pyproject.toml
[tool.poetry]
name = "python-activator"
version = "0.1.0"
description = ""
authors = ["Farid Seifi <farid.seifi@gmail.com>"]
readme = "README.md"
packages = [{include = "python_activator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.98.0"
flask = "^2.3.2"
pyyaml = "^6.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
uvicorn = "^0.22.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
python-activator="python_activator.cli:cli"


